Upsert cves for public images
